move apd new fields to components
